+++ docs/changelog.md
@@ -4,7 +4,8 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v6.0.0
 
-- \[API\] Rename some Animator APIs.
+- \[API\] Rename some Animator APIs
+- \[API\] Delete Animator.played() and advanced()
 - \[fix\] Infinite duration check for reverse
 
 ## v5.0.0